Update credential-bc-natural-gas-title-EO100.md
Hi, I revised this credential to completely follow the ToIP Metamodel. Any section that we don't have information yet or not applicable for this credential, I suggest to keep it there and input something like "This section is not applicable for EO 100 credential." This will prevent reader from being confused (to know whether the credential miss some part or the part is not applicable) in the future.