CH - Updated Tract Rights
Updated the Tract Rights section of the schema to articulate "included" and "excluded" to better describe this section. Also inputted feedback from yesterday related to how to articulate each piece of the Tract Section. This involves clearly laying out how each section appears on a title document (i.e. "there is a heading which is the tract number _________. Following the tract number there is a section called _____________" This is to clearly lay out the separate sections of the document.