Updated Credential based on meeting
Contact information is moved to the administering authority.

This document talks about the credential. In the purpose not good to say that it should be issued by emli this should be in the general or business requirements. (INPUTTED INTO BUSINESS REQUIREMENT)

The purpose of the governance framework is to outline all rules associated with governance issuance verification…… (this is now  included in the purpose.) (DONE)